--- src/components/navigationPanel/NavigationPanel.tsx
@@ -1,14 +1,0 @@
-import React from 'react';
-import { NavLink } from 'react-router-dom';
-
-function NavigationPanel() {
-  return (
-    <div className="nav-panel">
-      <NavLink to="/" className="nav-tab" exact>Home</NavLink>
-      <NavLink to="/new" className="nav-tab">New words</NavLink>
-      <NavLink to="/test" className="nav-tab">Test</NavLink>
-    </div>
-  );
-}
-
-export default NavigationPanel;
