+++ src/features/words/ListOfTrainings.scss
