+++ src/features/words/Training.scss
