--- src/scss/common.scss
@@ -1,4 +1,0 @@
-#root {
-  position: relative;
-  height: 100vh;
-}
