--- src/scss/main.scss
@@ -1,2 +1,0 @@
-@import './common.scss';
-@import '../components/navigationPanel/NavigationPanel.scss';
