+++ src/styles/main.scss
@@ -1,3 +1,8 @@
+#root {
+  position: relative;
+  height: 100vh;
+}
+
 .nav-panel {
   position: absolute;
   bottom: 0;
